+++ xmake.lua
@@ -2,9 +2,9 @@ add_rules("mode.debug", "mode.release")
 
 add_repositories("liteldev-repo https://github.com/LiteLDev/xmake-repo.git")
 
-add_requires("levilamina 1.1.0-rc.1", {configs = {target_type = "server"}})
+add_requires("levilamina 1.1.1", {configs = {target_type = "server"}})
 add_requires("levibuildscript 0.3.0")
-add_requires("legacyremotecall 0.9.0-rc.1")
+add_requires("legacyremotecall 0.10.0")
 
 if not has_config("vs_runtime") then
     set_runtimes("MD")
@@ -27,9 +27,7 @@ target("iListenAttentively-LseExport")
     add_defines(
         "NOMINMAX",
         "UNICODE",
-        "_HAS_CXX17",
-        "_HAS_CXX20",
-        "_HAS_CXX23"
+        "_HAS_CXX23=1"
     )
     add_packages(
         "levilamina",